shop: static buy menu

+++ src/state.h
@@ -10,7 +10,7 @@ const uint8_t EEPROM_START = 64;
 struct GameState {
     uint16_t minute;
     uint8_t baitCounts[4];
-    uint8_t money;
+    uint16_t money;
     bool acquiredFish[static_cast<int8_t>(FishType::COUNT)];
     uint8_t currentFishCount[static_cast<int8_t>(FishType::COUNT)];
 };
